parserAnalyzer: refactor folders
