+++ coverage.out
@@ -6,24 +6,24 @@ github.com/Dimix-international/in_memory_db-GO/db/db.go:27.45,35.2 5 1
 github.com/Dimix-international/in_memory_db-GO/db/db.go:37.33,44.2 4 1
 github.com/Dimix-international/in_memory_db-GO/db/db.go:46.42,53.2 4 1
 github.com/Dimix-international/in_memory_db-GO/db/db.go:55.42,60.2 3 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:21.26,23.2 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:25.56,26.23 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:26.23,27.18 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:28.19,29.46 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:29.46,31.5 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:32.4,33.38 2 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:34.33,35.31 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:35.31,39.10 4 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:41.4,41.46 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:41.46,43.5 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:44.4,44.28 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:45.24,46.31 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:46.31,47.13 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:49.4,49.46 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:49.46,51.5 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:53.4,54.38 2 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:58.2,58.41 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:58.41,60.3 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:62.2,62.22 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:65.38,67.2 1 1
-github.com/Dimix-international/in_memory_db-GO/internal/parser/parser.go:69.52,76.2 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:21.40,23.2 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:25.63,26.23 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:26.23,27.18 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:28.19,29.46 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:29.46,31.5 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:32.4,33.38 2 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:34.33,35.31 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:35.31,39.10 4 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:41.4,41.46 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:41.46,43.5 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:44.4,44.28 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:45.24,46.31 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:46.31,47.13 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:49.4,49.46 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:49.46,51.5 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:53.4,54.38 2 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:58.2,58.41 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:58.41,60.3 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:62.2,62.22 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:65.38,67.2 1 1
+github.com/Dimix-international/in_memory_db-GO/internal/service/parser.go:69.52,76.2 1 1
